+++ crates/proto/src/lib.rs
@@ -4,9 +4,15 @@ pub mod errors;
 #[rustfmt::skip]
 pub mod generated;
 
+pub mod clients;
+
 // RE-EXPORTS
 // ================================================================================================
 
+pub use clients::{
+    BlockProducerStoreClient, Client, ClientBuilder, ClientError, NtxBuilderStoreClient,
+    RpcStoreClient,
+};
 pub use domain::{
     account::{AccountState, AccountWitnessRecord},
     convert,